pyproject.toml
[tool.poetry]
name = "formee"
version = "0.1.0"
description = ""
authors = ["Arpan Pandey <arpanpandey.aps@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
PyInquirer = "^1.0.3"
pyfiglet = "^0.8.post1"
rich = "^12.0.1"
requests = "^2.27.1"
PyYAML = "^6.0"
gql = "^3.1.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
